requirements.txt
requests==2.23.0
pyyaml==5.3.1
